New subcommand: `query`  to support customized set of resources

+++ internal/test/resourcegroup/e2e_cases_test.go
@@ -2,21 +2,29 @@ package resourcegroup
 
 import (
 	"context"
+	"os"
+	"testing"
+	"time"
+
 	"github.com/Azure/aztfy/internal"
 	"github.com/Azure/aztfy/internal/config"
 	"github.com/Azure/aztfy/internal/test"
 	"github.com/Azure/aztfy/internal/test/cases"
+	"github.com/Azure/aztfy/internal/utils"
 	"github.com/hashicorp/terraform-exec/tfexec"
-	"os"
-	"path/filepath"
-	"testing"
 )
 
 func runCase(t *testing.T, d test.Data, c cases.Case) {
 	tfexecPath := test.EnsureTF(t)
+
 	provisionDir := t.TempDir()
+	if test.Keep() {
+		provisionDir, _ = os.MkdirTemp("", "")
+		t.Log(provisionDir)
+	}
+
 	os.Chdir(provisionDir)
-	if err := os.WriteFile("main.tf", []byte(c.Tpl(d)), 0644); err != nil {
+	if err := utils.WriteFileSync("main.tf", []byte(c.Tpl(d)), 0644); err != nil {
 		t.Fatalf("created to create the TF config file: %v", err)
 	}
 	tf, err := tfexec.NewTerraform(provisionDir, tfexecPath)
@@ -24,53 +32,50 @@ func runCase(t *testing.T, d test.Data, c cases.Case) {
 		t.Fatalf("failed to new terraform: %v", err)
 	}
 	ctx := context.Background()
+	t.Log("Running: terraform init")
 	if err := tf.Init(ctx); err != nil {
 		t.Fatalf("terraform init failed: %v", err)
 	}
+	t.Log("Running: terraform apply")
 	if err := tf.Apply(ctx); err != nil {
 		t.Fatalf("terraform apply failed: %v", err)
 	}
-	defer func() {
-		if err := tf.Destroy(ctx); err != nil {
-			t.Logf("terraform destroy failed: %v", err)
-		}
-	}()
+
+	if !test.Keep() {
+		defer func() {
+			t.Log("Running: terraform destroy")
+			if err := tf.Destroy(ctx); err != nil {
+				t.Logf("terraform destroy failed: %v", err)
+			}
+		}()
+	}
+
+	const delay = time.Minute
+	t.Logf("Sleep for %v to wait for the just created resources be recorded in ARG\n", delay)
+	time.Sleep(delay)
 
 	aztfyDir := t.TempDir()
 	resourceMapping, err := c.ResourceMapping(d)
 	if err != nil {
 		t.Fatalf("failed to get resource mapping: %v", err)
 	}
-	cfg := config.RgConfig{
+	cfg := config.GroupConfig{
 		CommonConfig: config.CommonConfig{
 			SubscriptionId: os.Getenv("ARM_SUBSCRIPTION_ID"),
 			OutputDir:      aztfyDir,
 			BackendType:    "local",
+			DevProvider:    true,
+			PlainUI:        true,
 		},
 		ResourceGroupName: d.RandomRgName(),
 		ResourceMapping:   resourceMapping,
+		RecursiveQuery:    true,
 	}
+	t.Logf("Batch importing the resource group %s\n", d.RandomRgName())
 	if err := internal.BatchImport(cfg, false); err != nil {
 		t.Fatalf("failed to run batch import: %v", err)
 	}
-	tf2, err := tfexec.NewTerraform(aztfyDir, tfexecPath)
-	if err != nil {
-		t.Fatalf("failed to new terraform: %v", err)
-	}
-	diff, err := tf2.Plan(ctx)
-	if err != nil {
-		t.Fatalf("terraform plan in the generated workspace failed: %v", err)
-	}
-	if diff {
-		t.Fatalf("terraform plan shows diff")
-	}
-	state, err := tf2.ShowStateFile(ctx, filepath.Join(aztfyDir, "terraform.tfstate"))
-	if err != nil {
-		t.Fatalf("terraform state show in the generated workspace failed: %v", err)
-	}
-	if n, expect := len(state.Values.RootModule.Resources), len(resourceMapping); n != expect {
-		t.Fatalf("expected terrafied resource: %d, got=%d", expect, n)
-	}
+	test.Verify(t, ctx, aztfyDir, tfexecPath, len(resourceMapping))
 }
 
 func TestComputeVMDisk(t *testing.T) {
